% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sktmap.r
\name{tmap_geocode}
\alias{tmap_geocode}
\title{R package for SK Telecom TMAP API
geocode  \code{\link{sktmap}}}
\usage{
tmap_geocode(addr, app_key)
}
\arguments{
\item{addr}{Target address character, which can be either old format (admin/lot number) or new format (road/building number)}

\item{app_key}{SK Telecom OPEN API perosonal key \url{http://tmapapi.sktelecom.com/main.html#web/guide/webGuide.sample1}}
}
\value{
A tibble of 26 varialbles is returned. Its elements depend on how the address is provided into the function; when "new" address template is used (road - builidng number), `newLat` and `newLon` are returned, while the "old" address template is used `lat` and `lon` are returned.
}
\description{
R package for SK Telecom TMAP API
geocode  \code{\link{sktmap}}
}
\author{
YD Hwang \email{yhwang@g.skku.edu}
}
