% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sktmap.r
\name{tmap_rev_geocode}
\alias{tmap_rev_geocode}
\title{reverse geocode function \code{\link{sktmap}}}
\usage{
tmap_rev_geocode(lat = 37.587228, lon = 126.993115, app_key)
}
\arguments{
\item{lat}{Target latitude. Default value is for SKKU in Seoul, Korea.}

\item{lon}{Target longitude. Default value is for SKKU in Seoul, Korea.}

\item{app_key}{SK Telecom OPEN API perosonal key from \href{http://tmapapi.sktelecom.com/main.html#web/guide/webGuide.sample1}{TMAP webpage.}}
}
\value{
A tibble of 18 varialbles is returned.
}
\description{
reverse geocode function \code{\link{sktmap}}
}
\details{
\code{addressType=A10} in the API gives the "full set" of the address; this is one of the options in TMAP API. In the package A10 is set as default. For the other options, check  \href{http://tmapapi.sktelecom.com/main.html#web/guide/webGuide.sample1}{TMAP webpage.}. This option provides the triple of:
\enumerate{
\item administrative district name (e.g., Sajikdong)
\item old address (admin-code and lot-number based)
\item new address (road-centric)
When returned, fullAddress are provided after separated accordingly; admin_name, jibun_address, and road_address.
}
}
\author{
YD Hwang \email{yhwang@g.skku.edu} and JK Kim \email{jkim@iastate.edu}
}
