% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sktmap.r
\name{tmap_route_search}
\alias{tmap_route_search}
\title{route search function \code{\link{sktmap}}}
\usage{
tmap_route_search(lat_origin, lon_origin, lat_dest, lon_dest, app_key)
}
\arguments{
\item{lat_origin}{origin's latitude.}

\item{lon_origin}{origin's longitude.}

\item{lat_dest}{destination latitude.}

\item{lon_dest}{destination longitude.}

\item{app_key}{SK Telecom OPEN API perosonal key \url{http://tmapapi.sktelecom.com/main.html#web/guide/webGuide.sample1}}
}
\value{
A Simple feature collection with 28 features and 16 fields. For more on simple feature in R, see sf package's manual.
}
\description{
route search function \code{\link{sktmap}}
}
\author{
YD Hwang \email{yhwang@g.skku.edu}
}
