% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sktmap.r
\name{tmap_rev_geocode}
\alias{tmap_rev_geocode}
\title{reverse geocode function \code{\link{sktmap}}}
\usage{
tmap_rev_geocode(lat = 37.587228, lon = 126.993115, app_key)
}
\arguments{
\item{lat}{Target latitude. Default value is for SKKU in Seoul, Korea.}

\item{lon}{Target longitude. Default value is for SKKU in Seoul, Korea.}

\item{app_key}{SK Telecom OPEN API perosonal key \url{http://tmapapi.sktelecom.com/main.html#web/guide/webGuide.sample1}}
}
\value{
A tibble of 18 varialbles is returned. addressType=A10: gives the "full set" of the address of the triple of (1) administrative district name, (2) old address (admin-code or lot-number based), (3) new address (road-centric). When returned, fullAddress are provided after separated accordingly; admin_name, jibun_address, and road_address.
}
\description{
reverse geocode function \code{\link{sktmap}}
}
\author{
YD Hwang \email{yhwang@g.skku.edu}
}
